+++ src/deprecated.jl
@@ -1,1 +1,3 @@
 
+Base.@deprecate_binding ManifoldGradientObjective ManifoldFirstOrderObjective
+Base.Base.@deprecate_binding AbstractManifoldGradientObjective AbstractManifoldFirstOrderObjective