+++ src/solvers/conjugate_gradient_descent.jl
@@ -57,7 +57,7 @@ the last direction ``δ_{k-1}`` and both gradients ``$(_tex(:grad))f(x_k)``,``$(
 The [`Stepsize`](@ref) ``s_k`` may be determined by a [`Linesearch`](@ref).
 
 Alternatively to `f` and `grad_f` you can provide
-the [`AbstractManifoldGradientObjective`](@ref) `gradient_objective` directly.
+the [`AbstractManifoldFirstOrderObjective`](@ref) `gradient_objective` directly.
 
 Available update rules are [`SteepestDescentCoefficientRule`](@ref), which yields a [`gradient_descent`](@ref),
 [`ConjugateDescentCoefficient`](@ref) (the default), [`DaiYuanCoefficientRule`](@ref), [`FletcherReevesCoefficient`](@ref),
@@ -88,7 +88,7 @@ $(_var(:Keyword, :stepsize; default="[`ArmijoLinesearch`](@ref)`()`"))
 $(_var(:Keyword, :stopping_criterion; default="[`StopAfterIteration`](@ref)`(500)`$(_sc(:Any))[`StopWhenGradientNormLess`](@ref)`(1e-8)`"))
 $(_var(:Keyword, :vector_transport_method))
 
-If you provide the [`ManifoldGradientObjective`](@ref) directly, the `evaluation=` keyword is ignored.
+If you provide the [`ManifoldFirstOrderObjective`](@ref) directly, the `evaluation=` keyword is ignored.
 The decorations are still applied to the objective.
 
 $(_note(:OutputSection))
@@ -105,13 +105,13 @@ function conjugate_gradient_descent(
     p_ = _ensure_mutating_variable(p)
     f_ = _ensure_mutating_cost(f, p)
     grad_f_ = _ensure_mutating_gradient(grad_f, p, evaluation)
-    mgo = ManifoldGradientObjective(f_, grad_f_; evaluation=evaluation)
+    mgo = ManifoldFirstOrderObjective(f_, grad_f_; evaluation=evaluation)
     rs = conjugate_gradient_descent(M, mgo, p_; evaluation=evaluation, kwargs...)
     return _ensure_matching_output(p, rs)
 end
 function conjugate_gradient_descent(
     M::AbstractManifold, mgo::O, p=rand(M); kwargs...
-) where {O<:Union{AbstractManifoldGradientObjective,AbstractDecoratedManifoldObjective}}
+) where {O<:Union{AbstractManifoldFirstOrderObjective,AbstractDecoratedManifoldObjective}}
     q = copy(M, p)
     return conjugate_gradient_descent!(M, mgo, q; kwargs...)
 end
@@ -126,7 +126,7 @@ function conjugate_gradient_descent!(
     evaluation::AbstractEvaluationType=AllocatingEvaluation(),
     kwargs...,
 ) where {TF,TDF}
-    mgo = ManifoldGradientObjective(f, grad_f; evaluation=evaluation)
+    mgo = ManifoldFirstOrderObjective(f, grad_f; evaluation=evaluation)
     dmgo = decorate_objective!(M, mgo; kwargs...)
     return conjugate_gradient_descent!(M, dmgo, p; kwargs...)
 end
@@ -144,7 +144,7 @@ function conjugate_gradient_descent!(
     vector_transport_method=default_vector_transport_method(M, typeof(p)),
     initial_gradient=zero_vector(M, p),
     kwargs...,
-) where {O<:Union{AbstractManifoldGradientObjective,AbstractDecoratedManifoldObjective}}
+) where {O<:Union{AbstractManifoldFirstOrderObjective,AbstractDecoratedManifoldObjective}}
     dmgo = decorate_objective!(M, mgo; kwargs...)
     dmp = DefaultManoptProblem(M, dmgo)
     cgs = ConjugateGradientDescentState(