+++ test/plans/test_problem.jl
@@ -5,12 +5,12 @@ using Manopt, Manifolds, Test
         M = Euclidean(3)
         f(M, p) = norm(p)^2
         grad_f(M, p) = 2 * p
-        moa = ManifoldGradientObjective(f, grad_f)
+        moa = ManifoldFirstOrderObjective(f, grad_f)
         cpa = DefaultManoptProblem(M, moa)
         @test Manopt.evaluation_type(cpa) === AllocatingEvaluation
 
         grad_f!(M, X, p) = (X .= 2 * p)
-        moi = ManifoldGradientObjective(f, grad_f!; evaluation=InplaceEvaluation())
+        moi = ManifoldFirstOrderObjective(f, grad_f!; evaluation=InplaceEvaluation())
         cpi = DefaultManoptProblem(M, moi)
         @test Manopt.evaluation_type(cpi) === InplaceEvaluation
     end